refactor: useForm 로직 개선
