+++ src/hooks/useAutoFocus/useAutoFocus.ts
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from 'react';
+
+export const useAutoFocus = () => {
+  const autoFocusRef = useRef<HTMLInputElement>(null);
+
+  useEffect(() => {
+    autoFocusRef.current?.focus();
+  }, []);
+
+  return autoFocusRef;
+};
